requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.1.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0

# Text processing and readability
textstat>=0.7.3
nltk>=3.8.1
spacy>=3.6.0

# Utilities
requests>=2.31.0
pyyaml>=6.0
click>=8.1.0

# Development and testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Additional utilities
validators>=0.20.0
psutil>=5.9.0
colorama>=0.4.6
